requirements.txt
asgiref==3.3.4
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
Django==3.2.4
djangorestframework==3.12.4
dynaconf==3.1.4
hvac==0.10.14
idna==2.10
loguru==0.5.3
Pillow==8.2.0
pip==21.1.2
pycparser==2.20
pytz==2021.1
requests==2.25.1
setuptools==52.0.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
validators==0.18.2
wheel==0.36.2
win32-setctime==1.0.3
wincertstore==0.2
